requirements.txt
langchain>=0.1.17
langchain-community>=0.0.17
langchain-openai>=0.1.3
python-dotenv
openai>=1.0.0
